tests: reuse db tests for mysql, postgres
