+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/db/Db_Common_Tests.kt
@@ -1,8 +1,15 @@
 package test.db
 
+import kiit.common.data.DataType
 import kiit.common.data.IDb
+import kiit.common.data.Value
+import kiit.common.ids.ULIDs
 import org.junit.Assert
 import org.junit.Test
+import test.entities.SampleEntityImmutable
+import test.setup.Address
+import test.setup.StatusEnum
+import java.util.*
 
 abstract class Db_Common_Tests : DbTestCases {
     abstract fun db(): IDb
@@ -23,12 +30,12 @@ abstract class Db_Common_Tests : DbTestCases {
         Assert.assertTrue(id > 0)
 
         // 2. update
-        val sqlUpdate = "update ${table()} set test_int = 987 where id = $id"
+        val sqlUpdate = "update ${table()} set test_int = 987 where ${encode("id")} = $id"
         val count = db.update(sqlUpdate)
         Assert.assertTrue(count > 0)
 
         // 3. get
-        val sql = "select test_int from ${table()} where id = $id"
+        val sql = "select test_int from ${table()} where ${encode("id")} = $id"
         val updatedVal = db.getScalarInt(sql, null)
         Assert.assertTrue(updatedVal == 987)
     }
@@ -41,7 +48,7 @@ abstract class Db_Common_Tests : DbTestCases {
         Assert.assertTrue(id > 0)
 
         // 2. get count
-        val sqlCount = "select count(*) from ${table()} where id = $id;"
+        val sqlCount = "select count(*) from ${table()} where ${encode("id")} = $id;"
         val countBefore = db.getScalarInt(sqlCount)
         Assert.assertTrue(countBefore == 1)
 
@@ -52,6 +59,41 @@ abstract class Db_Common_Tests : DbTestCases {
         // 4. get count after delete
         val countAfter = db.getScalarInt(sqlCount)
         Assert.assertTrue(countAfter == 0)
+    }
+
+
+    @Test
+    override fun can_get() {
+        val db = db()
+        // 1. add
+        val id = db.insert(insertSqlRaw())
+        Assert.assertTrue(id > 0)
+
+        // 2. update
+        val sqlGet = "select * from ${table()} where ${encode("id")} = $id;"
+        val item = db.mapOne(sqlGet, null) { rec ->
+            val longid = rec.getLong("id")
+            SampleEntityImmutable(
+                longid,
+                rec.getString("test_string"),
+                rec.getString("test_string_enc"),
+                rec.getBool("test_bool"),
+                rec.getShort("test_short"),
+                rec.getInt("test_int"),
+                rec.getLong("test_long"),
+                rec.getFloat("test_float"),
+                rec.getDouble("test_double"),
+                StatusEnum.convert(rec.getInt("test_enum")) as StatusEnum,
+                rec.getLocalDate("test_localdate"),
+                rec.getLocalTime("test_localtime"),
+                rec.getLocalDateTime("test_localdatetime"),
+                rec.getZonedDateTime("test_zoneddatetime"),
+                rec.getUUID("test_uuid"),
+                rec.getUPID("test_uniqueId"),
+                Address("", "", "", 1, "", true)
+            )
+        }
+        Assert.assertNotNull(item)
     }
 
 
@@ -126,4 +168,6 @@ abstract class Db_Common_Tests : DbTestCases {
 
     abstract fun table(): String
     abstract fun insertSqlRaw(): String
+    abstract fun insertSqlPrep(): String
+    abstract fun encode(column:String):String
 }
